requirements.txt
attrs==23.1.0
boto3==1.28.35
botocore==1.31.35
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
crashtest==0.4.1
distlib==0.3.7
dulwich==0.21.5
filelock==3.12.2
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jmespath==1.0.1
jsonschema==4.17.3
keyring==24.2.0
more-itertools==10.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
s3transfer==0.6.2
shellingham==1.5.3
six==1.16.0
tomlkit==0.12.1
trove-classifiers==2023.8.7
urllib3==1.26.16
virtualenv==20.24.3
xattr==0.10.1
zipp==3.16.2
